requirements.txt
six
Faker
pillow
markdown
python-dateutil
pytz
DateTime

django
django-environ
django-filter
django-allauth
django-resized
python-decouple
django-crispy-forms
requests

crispy-bootstrap5
django-browser-reload
drf-yasg
django-phonenumber-field[phonenumbers]

PyJWT
cryptography
paramiko

mailchimp_transactional
fido2